`size_t' instead of `unsigned long'

+++ C++/Configor.cpp
@@ -47,7 +47,7 @@ string CConfigorNode::GetName()
 	return m_Name;
 }
 
-unsigned long CConfigorNode::GetDataLength()
+size_t CConfigorNode::GetDataLength()
 {
 	return m_Length;
 }
@@ -64,7 +64,7 @@ void CConfigorNode::DestroyData()
 	m_pData = 0;
 }
 
-void CConfigorNode::SetData(unsigned char* pData, unsigned long Length)
+void CConfigorNode::SetData(unsigned char* pData, size_t Length)
 {
 	bool Destroy = Length > m_Length && Length - m_Length > 128; // Only re-allocate if we're more than 128 bytes
 
@@ -221,7 +221,7 @@ bool CConfigor::LoadFromFile(const std::string& Name)
 	}
 
 	ifs.seekg(0, ios::end);
-	unsigned long len = ifs.tellg();
+	size_t len = ifs.tellg();
 	ifs.seekg(0, ios::beg);
 
 	char* cnfg = new char[len];
@@ -239,7 +239,7 @@ bool CConfigor::LoadFromFile(const std::string& Name)
 
 
 // Returns error!
-string CConfigor::ParseQuotes(char** Out, unsigned long* Length, char* End)
+string CConfigor::ParseQuotes(char** Out, size_t* Length, char* End)
 {
 	string sb;
 
@@ -328,7 +328,7 @@ EOL:
 	return "";
 }
 
-CConfigorLexNodeList CConfigor::Lexify(char* Input, unsigned long Length)
+CConfigorLexNodeList CConfigor::Lexify(char* Input, size_t Length)
 {
 	CConfigorLexNodeList ret;
 	m_CurrentParseLine = 1;
@@ -345,7 +345,7 @@ CConfigorLexNodeList CConfigor::Lexify(char* Input, unsigned long Length)
 		case '"':
 			{
 				char* out;
-				unsigned long len;
+				size_t len;
 				string err = ParseQuotes(&out, &len, Input+Length);
 				if(err.length() > 1)
 				{
@@ -414,7 +414,7 @@ void PrintLexedNode(CConfigorLexNode* node)
 	}
 }
 
-bool CConfigor::LoadFromString(char* Input, unsigned long Length)
+bool CConfigor::LoadFromString(char* Input, size_t Length)
 {
 	CConfigorLexNodeList lexed = Lexify(Input, Length);
 
@@ -454,7 +454,7 @@ bool CConfigor::LoadFromString(char* Input, unsigned long Length)
 				it2++;
 
 				char* pName = (char*)node->Value;
-				unsigned long mLength = node->Length;
+				size_t mLength = node->Length;
 
 				IConfigorNode* node = new CConfigorNode(pCurrentNode, string(pName, mLength));
 
@@ -464,7 +464,7 @@ bool CConfigor::LoadFromString(char* Input, unsigned long Length)
 				{
 					it++;
 					unsigned char* pData = (unsigned char*)(*it2)->Value;
-					unsigned long dLength = (*it2)->Length;
+					size_t dLength = (*it2)->Length;
 					node->SetData(pData, dLength);
 					if(dLength)
 						delete [] pData;
@@ -539,12 +539,12 @@ bool ValidChar(char x)
 }
 
 // Takes data + length, outputs char + length
-void EscapeData(unsigned char* pData, unsigned long Length, unsigned char** Output, unsigned long* OutLength) 
+void EscapeData(unsigned char* pData, size_t Length, unsigned char** Output, size_t* OutLength) 
 {
 	string sb;
 	sb.reserve(Length);
 
-	for(unsigned long i = 0; i < Length; i++)
+	for(size_t i = 0; i < Length; i++)
 	{
 		char x = (char)pData[i];
 
@@ -595,7 +595,7 @@ void EscapeData(unsigned char* pData, unsigned long Length, unsigned char** Outp
 void RecursiveSave(ostream& ofs, IConfigorNode* node, int Depth = 0)
 {
 	unsigned char* pData;
-	unsigned long Length;
+	size_t Length;
 
 	EscapeData((unsigned char*)node->GetName().c_str(), node->GetName().size(), &pData, &Length);
